% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_diff_tree}
\alias{git_diff_tree}
\title{Get the difference between two tree-ish}
\usage{
git_diff_tree(left, right, repo = NULL)
}
\arguments{
\item{left}{the tree-ish to use as the base for the comparison.}

\item{right}{the tree-ish to compare to the base.}

\item{repo}{the path to the Git repo to use.}
}
\value{
a dataframe for each differing entry in the trees, with columns
\code{mode1}, \code{mode2}, \code{hash1}, \code{hash2}, \code{status}, \code{score}, \code{src} and \code{dst}.
}
\description{
Get the difference between two tree-ish
}
