% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_get_latest_commit}
\alias{git_get_latest_commit}
\title{Get the last commit to have modified the given path.}
\usage{
git_get_latest_commit(path, ref, repo = NULL)
}
\arguments{
\item{path}{path to the file of directory to search for}

\item{ref}{the Git commit from which to start the search. Only ancestors of
that commit will be considered.}

\item{repo}{the path to the Git repo to use.}
}
\description{
If the path is a directory, any modification to files contained within
it are considered.
}
