% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{Queue}
\alias{Queue}
\title{Object for managing running jobs on the redis queue}
\description{
Object for managing running jobs on the redis queue

Object for managing running jobs on the redis queue
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{root}}{Orderly root}

\item{\code{config}}{Orderly config}

\item{\code{controller}}{RRQ controller}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Queue-new}{\code{Queue$new()}}
\item \href{#method-Queue-number_of_workers}{\code{Queue$number_of_workers()}}
\item \href{#method-Queue-finalize}{\code{Queue$finalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-new"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-new}{}}}
\subsection{Method \code{new()}}{
Create object, read configuration and setup Redis connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$new(root, queue_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{root}}{Orderly root.}

\item{\code{queue_id}}{ID of an existing queue to connect to, creates a new one
if NULL (default NULL)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-number_of_workers"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-number_of_workers}{}}}
\subsection{Method \code{number_of_workers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$number_of_workers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-finalize}{}}}
\subsection{Method \code{finalize()}}{
Destroy queue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$finalize()}\if{html}{\out{</div>}}
}

}
}
