% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{repository_path}
\alias{repository_path}
\title{Get the storage path for a repository.}
\usage{
repository_path(base, url, check = TRUE)
}
\arguments{
\item{base}{the base directory in which all repositories are stored.}

\item{url}{the URL of the remote repository.}

\item{check}{if TRUE and the repository does not exist locally yet, an error
is raised.}
}
\value{
the path to the repository.
}
\description{
Repositories are all stored as subdirectories of a common base path,
using a hash of their URL as the directory name.
}
