% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Create orderly runner}
\usage{
api(
  repositories_base_path,
  validate = NULL,
  log_level = "info",
  skip_queue_creation = FALSE
)
}
\arguments{
\item{repositories_base_path}{Path in which Git repositories are
cloned.}

\item{validate}{Logical, indicating if validation should be done
on responses.  This should be \code{FALSE} in production
environments.  See \link[porcelain:porcelain]{porcelain::porcelain} for details}

\item{log_level}{Logging level to use. Sensible options are "off",
"info" and "all".}

\item{skip_queue_creation}{Skip queue creation, this is primarily
used for tests where we can't establish a Redis connection.}
}
\value{
A \link[porcelain:porcelain]{porcelain::porcelain} object. Notably this does \emph{not}
start the server
}
\description{
Create an orderly runner, a porcelain object
}
