% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{Queue}
\alias{Queue}
\title{Object for managing running jobs on the redis queue}
\description{
Object for managing running jobs on the redis queue

Object for managing running jobs on the redis queue
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{root}}{Orderly root}

\item{\code{config}}{Orderly config}

\item{\code{controller}}{RRQ controller}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Queue-new}{\code{Queue$new()}}
\item \href{#method-Queue-submit}{\code{Queue$submit()}}
\item \href{#method-Queue-number_of_workers}{\code{Queue$number_of_workers()}}
\item \href{#method-Queue-get_status}{\code{Queue$get_status()}}
\item \href{#method-Queue-finalize}{\code{Queue$finalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-new"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-new}{}}}
\subsection{Method \code{new()}}{
Create object, read configuration and setup Redis connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$new(root, queue_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{root}}{Orderly root.}

\item{\code{queue_id}}{ID of an existing queue to connect to, creates a new one
if NULL (default NULL)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-submit"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-submit}{}}}
\subsection{Method \code{submit()}}{
Submit a job the Redis queue for runner to run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$submit(reportname, parameters = NULL, branch = "master", ref = "HEAD")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reportname}}{Name of orderly report.}

\item{\code{parameters}}{Parameters to run the report with (default NULL)}

\item{\code{branch}}{Name of git branch to checkout the repository
(default master)}

\item{\code{ref}}{Git commit-ish value (e.g HEAD or commit hash or branch name).
We reset hard to this ref and run the report. (default HEAD)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-number_of_workers"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-number_of_workers}{}}}
\subsection{Method \code{number_of_workers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$number_of_workers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-get_status}{}}}
\subsection{Method \code{get_status()}}{
Gets status of packet run
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$get_status(job_ids, include_logs = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job_ids}}{Id of redis queue job to get status of.}

\item{\code{include_logs}}{Whether to include logs in response or not}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
status of redis queue job
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Queue-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-Queue-finalize}{}}}
\subsection{Method \code{finalize()}}{
Destroy queue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$finalize()}\if{html}{\out{</div>}}
}

}
}
